moved more files into folder structures
