--- src/server/tests/sample.test.js
@@ -1,7 +1,0 @@
-/* global describe, it, expect */
-
-describe("Sample Test", () => {
-  it("should test that true === true", () => {
-    expect(true).toBe(true);
-  });
-});
