+++ src/server/utils/functions.test.ts
@@ -1,4 +1,4 @@
-import { fieldPeriodToText } from "../Functions";
+import { fieldPeriodToText } from "./functions";
 
 describe("Field period to text test", () => {
   it("should return 'April 2020' for field period 2004", () => {
